Revert "[Test] MemberAuthService"
