+++ src/main/java/com/todorian/_core/error/GlobalExceptionHandler.java
@@ -1,16 +1,9 @@
 package com.todorian._core.error;
 
 import com.todorian._core.error.exception.*;
-import com.todorian._core.utils.ApiUtils;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.util.HashMap;
-import java.util.Map;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
@@ -38,18 +31,5 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(Exception500.class)
     public ResponseEntity<?> serverError(Exception500 e){
         return new ResponseEntity<>(e.body(), e.status());
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<?> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-        for (FieldError error : ex.getBindingResult().getFieldErrors()) {
-            errors.put(error.getField(), error.getDefaultMessage());
-        }
-
-        // errors 맵의 첫 번째 요소의 값을 가져옴
-        String firstErrorMessage = errors.values().stream().findFirst().orElse("유효성 검사 오류가 발생했습니다.");
-
-        return new ResponseEntity<>(ApiUtils.error(firstErrorMessage, HttpStatus.BAD_REQUEST), HttpStatus.BAD_REQUEST);
     }
 }
