--- src/main/java/com/todorian/character/command/domain/model/CharacterCategory.java
@@ -1,4 +1,0 @@
-package com.todorian.character.command.domain.model;
-
-public enum CharacterCategory {
-}
