--- src/main/java/com/todorian/member/command/application/controller/MemberAuthController.java
@@ -1,77 +1,0 @@
-package com.todorian.member.command.application.controller;
-
-import com.todorian._core.utils.ApiUtils;
-import com.todorian.member.command.application.dto.MemberRequestDTO;
-import com.todorian.member.command.application.dto.MemberResponseDTO;
-import com.todorian.member.command.application.service.MemberAuthService;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api/auth")
-public class MemberAuthController {
-
-    private final MemberAuthService memberAuthService;
-
-    /*
-        기본 회원 가입
-     */
-    @PostMapping("/signup")
-    public ResponseEntity<?> signUp(@Valid @RequestBody MemberRequestDTO.signUpDTO requestDTO) {
-
-        memberAuthService.signUp(requestDTO);
-
-        return ResponseEntity.ok().body(ApiUtils.success(null));
-    }
-
-    /*
-        기본 로그인
-     */
-    @PostMapping("/login")
-    public ResponseEntity<?> login(HttpServletRequest httpServletRequest, @Valid @RequestBody MemberRequestDTO.loginDTO requestDTO) {
-
-        MemberResponseDTO.authTokenDTO responseDTO = memberAuthService.login(httpServletRequest, requestDTO);
-
-        return ResponseEntity.ok()
-                .header(HttpHeaders.AUTHORIZATION, responseDTO.grantType() + " " + responseDTO.accessToken())
-                .header("Refresh-Token", responseDTO.grantType() + " " + responseDTO.refreshToken())
-                .body(ApiUtils.success(null));
-    }
-
-    /*
-        Access Token 재발급 - Refresh Token 필요
-     */
-    @PostMapping("/reissue")
-    public ResponseEntity<?> reissueToken(HttpServletRequest httpServletRequest) {
-
-        MemberResponseDTO.authTokenDTO responseDTO = memberAuthService.reissueToken(httpServletRequest);
-
-        return ResponseEntity.ok()
-                .header(HttpHeaders.AUTHORIZATION, responseDTO.grantType() + " " + responseDTO.accessToken())
-                .header("Refresh-Token", responseDTO.grantType() + " " + responseDTO.refreshToken())
-                .body(ApiUtils.success(null));
-    }
-
-    /*
-        로그아웃 - Refresh Token 필요
-     */
-    @PostMapping("/logout")
-    public ResponseEntity<?> logout(HttpServletRequest httpServletRequest) {
-
-        log.info("로그아웃 시도");
-
-        memberAuthService.logout(httpServletRequest);
-
-        return ResponseEntity.ok().body(ApiUtils.success(null));
-    }
-}
