--- src/main/java/com/todorian/member/command/domain/model/Authority.java
@@ -1,5 +1,0 @@
-package com.todorian.member.command.domain.model;
-
-public enum Authority {
-    ADMIN, USER, GUEST, NONE
-}
