--- src/main/java/com/todorian/member/command/domain/model/SocialType.java
@@ -1,8 +1,0 @@
-package com.todorian.member.command.domain.model;
-
-public enum SocialType {
-    GOOGLE,
-    KAKAO,
-    NAVER,
-    NONE
-}
