--- src/main/java/com/todorian/member/command/domain/model/Status.java
@@ -1,6 +1,0 @@
-package com.todorian.member.command.domain.model;
-
-public enum Status {
-    // 활성, 휴면, 탈퇴, 정지
-    ACTIVE, DORMANT, DEACTIVATED, SUSPENDED
-}
