--- src/main/java/com/todorian/member/command/domain/model/property/KakaoProperties.java
@@ -1,16 +1,0 @@
-package com.todorian.member.command.domain.model.property;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Data
-@Configuration
-@ConfigurationProperties(prefix = "kakao")
-public class KakaoProperties {
-    private String tokenUri;
-    private String userInfoUri;
-    private String grantType;
-    private String clientId;
-    private String redirectUri;
-}
