--- src/main/java/com/todorian/member/command/domain/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.todorian.member.command.domain.repository;
-
-import com.todorian.member.command.domain.model.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
-
-}
