--- src/main/java/com/todorian/member/command/domain/repository/MemberRepositoryCustom.java
@@ -1,10 +1,0 @@
-package com.todorian.member.command.domain.repository;
-
-import com.todorian.member.command.domain.model.Member;
-
-import java.util.Optional;
-
-public interface MemberRepositoryCustom {
-
-    Optional<Member> findByEmail(String email);
-}
