--- src/main/java/com/todorian/member/command/domain/service/MemberOAuthService.java
@@ -1,9 +1,0 @@
-package com.todorian.member.command.domain.service;
-
-
-import com.todorian.member.command.application.dto.MemberResponseDTO;
-
-public interface MemberOAuthService {
-
-    MemberResponseDTO.authTokenDTO kakaoLogin(String code);
-}
