+++ src/main/java/com/todorian/member/domain/Member.java
@@ -1,0 +1,30 @@
+package com.todorian.member.domain;
+
+import com.todorian.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "member_tb")
+public class Member extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length = 100, nullable = false, unique = true)
+    private String email;
+    @Column(length = 100, nullable = false, unique = true)
+    private String password;
+
+    @Builder
+    public Member(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+}
