+++ src/main/java/com/todorian/member/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.todorian.member.repository;
+
+import com.todorian.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+
+}
