--- src/main/java/com/todorian/membercharacter/command/domain/service/MemberCharacterService.java
@@ -1,45 +1,0 @@
-package com.todorian.membercharacter.command.domain.service;
-
-import com.todorian.membercharacter.command.domain.model.MemberCharacter;
-import com.todorian.membercharacter.command.domain.model.MemberCharacterCreateRequestDTO;
-import com.todorian.membercharacter.command.domain.repository.MemberCharacterRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-public class MemberCharacterService {
-
-    @Autowired
-    private MemberCharacterRepository memberCharacterRepository;
-
-    public MemberCharacterService(MemberCharacterRepository memberCharacterRepository) {
-        this.memberCharacterRepository = memberCharacterRepository;
-    }
-
-    public void save(MemberCharacterCreateRequestDTO memberCharacterInfo) {
-        MemberCharacter memberCharacter = new MemberCharacter(
-                memberCharacterInfo.getMemberId(),
-                memberCharacterInfo.getCharacterId(),
-                memberCharacterInfo.getAcquisitionTime()
-        );
-
-        memberCharacterRepository.save(memberCharacter);
-    }
-
-    public void delete(long memberCharacterId) {
-        memberCharacterRepository.deleteById(memberCharacterId);
-    }
-
-    public List<MemberCharacter> findAll(){
-        List<MemberCharacter> memberCharacterList = memberCharacterRepository.findAll();
-
-        return memberCharacterList;
-    }
-
-    public MemberCharacter findById(long memberCharacterId) {
-        return memberCharacterRepository.findById(memberCharacterId).get();
-    }
-}
