--- src/main/java/com/todorian/redis/repository/RefreshTokenRedisRepository.java
@@ -1,9 +1,0 @@
-package com.todorian.redis.repository;
-
-import com.todorian.redis.domain.RefreshToken;
-import org.springframework.data.repository.CrudRepository;
-
-public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, Long> {
-
-    RefreshToken findByRefreshToken(String refreshToken);
-}
