--- src/main/resources/application.properties
@@ -1,1 +1,0 @@
-JWT.SECRET=8704b6707c11f6225d354c624a41212651b1c8357ef609ecb42688d3a25ac4bdea3b92accfd5acc4108a53eb44f36caf0941c80616967e503e17e251ef078ba9
