+++ src/main/resources/application.yml
@@ -1,4 +1,22 @@
 spring:
-  profiles:
-    active:
-      - local
+  datasource:
+    # url: jdbc:h2:tcp://localhost/~/test;MODE=MariaDB
+    url: jdbc:h2:mem:test;MODE=MariaDB
+    driver-class-name: org.h2.Driver
+    username: sa
+    password:
+  h2:
+    console:
+      enabled: true
+
+  application:
+    name: todorian
+
+  jpa:
+    show-sql: true
+    database: mysql
+    hibernate:
+      ddl-auto: create
+    properties:
+      hibernate:
+        format_sql: true
