--- src/main/resources/config/application-local.yml
@@ -1,29 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:h2:mem:test;MODE=MySQL
-    driver-class-name: org.h2.Driver
-    username: sa
-    password:
-  h2:
-    console:
-      enabled: true
-
-  application:
-    name: todorian
-
-  jpa:
-    show-sql: true
-    database: mysql
-    hibernate:
-      ddl-auto: create
-    properties:
-      hibernate:
-        format_sql: true
-
-jwt:
-  secret: ${JWT.SECRET}
-
-logging:
-  level:
-    '[com.todorian]': debug
-    '[org.hibernate.type]': trace
