--- src/main/resources/log/console-appender.xml
@@ -1,7 +1,0 @@
-<included>
-    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>${CONSOLE_LOG_PATTERN}</pattern>
-        </encoder>
-    </appender>
-</included>
