--- src/main/resources/logback-spring.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<configuration>
-    <timestamp key="BY_DATE" datePattern="yyyy-MM-dd"/>
-    <property name="CONSOLE_LOG_PATTERN"
-              value="[%d{yyyy-MM-dd HH:mm:ss}:%-4relative] %green([%thread]) %highlight(%-5level) %boldWhite([%C.%M:%yellow(%L)]) - %msg%n"/>
-    <property name="FILE_LOG_PATTERN"
-              value="[%d{yyyy-MM-dd HH:mm:ss}:%-4relative] [%thread] %-5level [%C.%M:%L] - %msg%n"/>
-
-    <springProfile name="!prod">
-        <include resource="log/console-appender.xml"/>
-
-        <logger name="com.todorian" level="DEBUG" />
-        <root level="INFO">
-            <appender-ref ref="CONSOLE"/>
-        </root>
-    </springProfile>
-
-    <springProfile name="prod">
-        <include resource="log/file-info-appender.xml"/>
-        <include resource="log/file-warn-appender.xml"/>
-        <include resource="log/file-error-appender.xml"/>
-
-        <root level="INFO">
-            <appender-ref ref="FILE-INFO"/>
-            <appender-ref ref="FILE-WARN"/>
-            <appender-ref ref="FILE-ERROR"/>
-        </root>
-    </springProfile>
-</configuration>
